See description below
I think im basically done with the printActionItem screen. It has the name of the action item, priority, comment, all the dates that are not null, and its history. it also has a print button at the bottom. The only thing that isn't working is the print button (i just have it to just system.out.println("print") and that the print button is too big. If someone can fix those, it would be appreciated.